djackets_django/requirements.txt
Django==3.2.8
Pillow==8.4.0
